chore: upgrade to rails 6.0
