+++ app/models/concerns/has_dns_checks.rb
@@ -1,6 +1,6 @@
 require "resolv"
 
-class Domain
+module HasDNSChecks
 
   def dns_ok?
     spf_status == "OK" && dkim_status == "OK" && ["OK", "Missing"].include?(mx_status) && ["OK", "Missing"].include?(return_path_status)
