+++ app/models/user.rb
@@ -27,11 +27,11 @@ class User < ApplicationRecord
 
   include HasUUID
 
-  require_dependency "user/authentication"
+  include HasAuthentication
 
   validates :first_name, presence: true
   validates :last_name, presence: true
-  validates :email_address, presence: true, uniqueness: true, format: { with: /@/, allow_blank: true }
+  validates :email_address, presence: true, uniqueness: { case_sensitive: false }, format: { with: /@/, allow_blank: true }
   validates :time_zone, presence: true
 
   default_value :time_zone, -> { "UTC" }
