+++ app/views/layouts/application.html.haml
@@ -58,4 +58,3 @@
               Powered by #{link_to "Postal", "https://postalserver.io", target: '_blank'} #{Postal.version}.
             %li= link_to "Documentation", "https://docs.postalserver.io", target: '_blank'
             %li= link_to "Ask for help", "https://discussions.postalserver.io", target: '_blank'
-
